[Refactor] (code + test) Rename 'cluster' module to be 'crawlers'.
