+++ smoothcrawler_cluster/crawler/__init__.py
@@ -1,1 +1,1 @@
-from .cluster import ZookeeperCrawler
+from .crawlers import ZookeeperCrawler
