+++ test/unit_test/crawler/cluster.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, Mock, patch
 import pytest
 from kazoo.client import KazooClient
 
-from smoothcrawler_cluster.crawler.cluster import ZookeeperCrawler
+from smoothcrawler_cluster.crawler.crawlers import ZookeeperCrawler
 from smoothcrawler_cluster.crawler.workflow import (
     BaseRoleWorkflow,
     PrimaryBackupRunnerWorkflow,
